Create README.md
I'm Harshita. I currently pursuing BE in CSE since 2022. When I coded for the first time and by seeing the relevant output to my code I was excited to explore more. I like to get inspired by the professionals especially in the IT field. Painting is my hobby.
My short term goal is to get an internship in my preferred domain.